requirements.txt
sqlite-database
pywebview
tqdm
requests
argh